Feat/field oriented drive (#57)
* add field oriented drive

* fix the limiting of drive speed to steer error

* fix the gyro readings (read the navx docs, had to to the Omnimount thing)

* actual field oriented drive

* fix field-oriented drive, and thus far it works pretty well (not entirely well, but that's more a matter for feat/swerve-drive-tuning or something)

* put the inversion logic for teleop back in

* second driver camera

* make default speed and acceleration faster

* add drive deadband, remove duplicated code

+++ src/main/include/io/SwerveDriveIO.h
@@ -59,7 +59,7 @@ typedef struct swerve_drive_hardware_interface_t {
 typedef struct swerve_drive_software_interface_t {
     // SW INPUTS
     // IMU yaw angle
-    double imu_yaw;
+    float imu_yaw;
     float imu_pitch;
 
     // Encoders
